Set exit code and log error, instead of throwing
Throwing an error when linked packages are found causes a
stracktrace to be written to output, which makes it look
like the script somehow failed.

Using console.error() to write the message, and setting
the exit code to 1 achieve the same result, but
without the ugly and somewhat confusing stacktrac the same result.